requirements.txt
attrs==24.2.0
awscli==1.34.10
backports.tarfile==1.2.0
bitarray==2.9.2
bitstring==4.2.3
boto3==1.35.10
botocore==1.35.10
brevitas==0.10.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clize==5.0.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
dependencies==2.0.1
dill==0.3.8
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
humanfriendly==10.0
idna==3.8
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.4
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
od==2.0.2
onnx==1.16.2
onnxruntime==1.19.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
psutil==6.0.0
pyasn1==0.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qonnx==0.3.0
requests==2.32.3
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
sigtools==4.0.1
six==1.16.0
sympy==1.13.2
tabulate==0.9.0
toposort==1.10
torch==1.3.1
torchvision==0.14.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
unfoldNd==0.2.2
urllib3==2.2.2
voluptuous==0.15.2
websocket-client==1.8.0
yolo==0.3.2
